REF (feedback): move dict_compat to core/dtypes/cast.py
